# src/main/resources/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Contact API",
        "description": "An API that can manage contacts",
        "version": "v1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "paths": {
        "/contact": {
            "post": {
                "tags": [
                    "Section Three"
                ],
                "operationId": "createContact",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Contact"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Contact"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/contact/{id}": {
            "get": {
                "tags": [
                    "Serction Two"
                ],
                "operationId": "getContact",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Contact"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/contact/all": {
            "get": {
                "tags": [
                    "Section One"
                ],
                "operationId": "getContacts",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Contact"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Contact": {
                "required": [
                    "name",
                    "phoneNumber"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "phoneNumber": {
                        "type": "string"
                    }
                }
            }
        }
    }
}